app/crypto-agent: fix RSA modlen type
RSA modlen type needs to be uint16_t in order to support larger
modulus value.

+++ app/crypto-agent/ca_crypto_queue.h
@@ -21,7 +21,7 @@ struct __rte_aligned(ROC_ALIGN) cpt_inflight_req
 	union dao_cpt_res_s res;
 	struct rte_mbuf *mbuf;
 	uint8_t rsa_is_decrypt : 1;
-	uint8_t rsa_mod_len;
+	uint16_t rsa_mod_len;
 };
 
 DAO_STATIC_ASSERT(sizeof(struct cpt_inflight_req) == 128);
